+++ DH2650/Assets/Materials/Shaders/CaveMaterial.mat
@@ -53,11 +53,11 @@ Material:
     - Vector1_5a3548f1dfa94117a7f1e56c30d9a830: 0.18
     - Vector1_795200d3bf8343ed9a802b8f95d40b22: 0.13
     - Vector1_e1e7828ab6a346428d0a250ad43a709f: 0.1
-    - Vector1_f7c20e0339a444e9a27f1790f618dd77: 0.52
-    - Vector1_fc2bddfbb3844fd38f2dbeebde518ef9: 0.29
+    - Vector1_f7c20e0339a444e9a27f1790f618dd77: 0.09
+    - Vector1_fc2bddfbb3844fd38f2dbeebde518ef9: 0.27
     m_Colors:
-    - Color_15a87427b61f476ab1c034a556d85094: {r: 0.33962262, g: 0.33161265, b: 0.33161265, a: 0}
-    - Color_6d7a49c753a14b73ada18210cb5f3176: {r: 0.4811321, g: 0.4811321, b: 0.4811321, a: 0}
+    - Color_15a87427b61f476ab1c034a556d85094: {r: 0.40616736, g: 0.3910644, b: 0.49056602, a: 0}
+    - Color_6d7a49c753a14b73ada18210cb5f3176: {r: 0.3018868, g: 0.3018868, b: 0.3018868, a: 0}
     - Vector2_07accd06c947428597015c04eb7c92a0: {r: 2, g: 2, b: 0, a: 0}
     - Vector2_1e396deee9114266ac8eee5c59a6b75c: {r: 1, g: 1, b: 0, a: 0}
     - Vector2_4ca26172cca7446d822604a639332105: {r: 0, g: 0, b: 0, a: 0}
