--- DH2650/Assets/Prefabs/Terrain/Trees/Palm_Tree Variant.prefab
@@ -1,90 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!64 &2737926935720997099
-MeshCollider:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 562837224788576042}
-  m_Material: {fileID: 0}
-  m_IsTrigger: 0
-  m_Enabled: 1
-  serializedVersion: 4
-  m_Convex: 0
-  m_CookingOptions: 30
-  m_Mesh: {fileID: -5220702601083036729, guid: a164b1db21026408d9485bd5648170c9, type: 3}
---- !u!1001 &796838791488666235
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: -8679921383154817045, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_RootOrder
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: -36.40179
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 319.34473
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: -11.038809
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0.000000021855694
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -8679921383154817045, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: -7511558181221131132, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_Materials.Array.data[0]
-      value: 
-      objectReference: {fileID: 4874126678125670354, guid: 83cf8183affd641279dd083d80019b41, type: 3}
-    - target: {fileID: -7511558181221131132, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_Materials.Array.data[1]
-      value: 
-      objectReference: {fileID: -6478536213147159724, guid: 83cf8183affd641279dd083d80019b41, type: 3}
-    - target: {fileID: 919132149155446097, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_Name
-      value: Palm_Tree
-      objectReference: {fileID: 0}
-    - target: {fileID: 919132149155446097, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-      propertyPath: m_Layer
-      value: 6
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: a164b1db21026408d9485bd5648170c9, type: 3}
---- !u!1 &562837224788576042 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: a164b1db21026408d9485bd5648170c9, type: 3}
-  m_PrefabInstance: {fileID: 796838791488666235}
-  m_PrefabAsset: {fileID: 0}
