--- DH2650/Assets/Prefabs/Terrain/Trees/Palm_Tree Variant.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b9e2db1fd30cf3f40aadc39686f2cf6f
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
