+++ DH2650/Assets/Prefabs/Terrain/Trees/Palms_4 Variant.prefab
@@ -14,20 +14,6 @@ MeshCollider:
   m_Convex: 0
   m_CookingOptions: 30
   m_Mesh: {fileID: -1682125421048621038, guid: a7678edfaae744811a37d52853df3540, type: 3}
---- !u!64 &7703008014921319551
-MeshCollider:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 9126060144634605336}
-  m_Material: {fileID: 0}
-  m_IsTrigger: 0
-  m_Enabled: 1
-  serializedVersion: 4
-  m_Convex: 0
-  m_CookingOptions: 30
-  m_Mesh: {fileID: -6062578164958778753, guid: a7678edfaae744811a37d52853df3540, type: 3}
 --- !u!1001 &702732845282622718
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -93,17 +79,12 @@ PrefabInstance:
       objectReference: {fileID: 4874126678125670354, guid: a7678edfaae744811a37d52853df3540, type: 3}
     - target: {fileID: 8603806674150118374, guid: a7678edfaae744811a37d52853df3540, type: 3}
       propertyPath: m_Layer
-      value: 6
+      value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: a7678edfaae744811a37d52853df3540, type: 3}
 --- !u!1 &360835652794499503 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: a7678edfaae744811a37d52853df3540, type: 3}
-  m_PrefabInstance: {fileID: 702732845282622718}
-  m_PrefabAsset: {fileID: 0}
---- !u!1 &9126060144634605336 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 8603806674150118374, guid: a7678edfaae744811a37d52853df3540, type: 3}
   m_PrefabInstance: {fileID: 702732845282622718}
   m_PrefabAsset: {fileID: 0}
