+++ DH2650/Assets/Prefabs/Terrain/Trees/Palms_5 Variant.prefab
@@ -14,20 +14,6 @@ MeshCollider:
   m_Convex: 0
   m_CookingOptions: 30
   m_Mesh: {fileID: -1682125421048621038, guid: 40b7c255c2d5046518754f7da42b6e2c, type: 3}
---- !u!64 &5083041176762472357
-MeshCollider:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6033296812125590366}
-  m_Material: {fileID: 0}
-  m_IsTrigger: 0
-  m_Enabled: 1
-  serializedVersion: 4
-  m_Convex: 0
-  m_CookingOptions: 30
-  m_Mesh: {fileID: -6062578164958778753, guid: 40b7c255c2d5046518754f7da42b6e2c, type: 3}
 --- !u!1001 &2656085745875566776
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -97,17 +83,12 @@ PrefabInstance:
       objectReference: {fileID: 4874126678125670354, guid: 83cf8183affd641279dd083d80019b41, type: 3}
     - target: {fileID: 8603806674150118374, guid: 40b7c255c2d5046518754f7da42b6e2c, type: 3}
       propertyPath: m_Layer
-      value: 6
+      value: 0
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 40b7c255c2d5046518754f7da42b6e2c, type: 3}
 --- !u!1 &2890508558538472937 stripped
 GameObject:
   m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: 40b7c255c2d5046518754f7da42b6e2c, type: 3}
-  m_PrefabInstance: {fileID: 2656085745875566776}
-  m_PrefabAsset: {fileID: 0}
---- !u!1 &6033296812125590366 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 8603806674150118374, guid: 40b7c255c2d5046518754f7da42b6e2c, type: 3}
   m_PrefabInstance: {fileID: 2656085745875566776}
   m_PrefabAsset: {fileID: 0}
