+++ DH2650/Assets/Scripts/Cave/SetCaveLighning.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SetCaveLighning : MonoBehaviour
+{
+
+    public GameObject dirSkyLight;
+
+    private void OnTriggerEnter(Collider other)
+    {
+        dirSkyLight.GetComponent<Light>().enabled = false;
+    }
+
+
+}
