+++ DH2650/Assets/Scripts/Cave/SetCaveLighning.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8b40aecabb13ed141b4a03bebf03da9e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
