--- DH2650/Assets/Scripts/DisableWater.cs
@@ -1,21 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class DisableWater : MonoBehaviour
-{
-    public GameObject water;
-    public GameObject waterContainer;
-
-    private void OnTriggerEnter(Collider other)
-    {
-        print("Disable!");
-        water.GetComponent<waves>().enabled = false;
-        water.GetComponent<WaterEffects>().enabled = false;
-        water.GetComponent<MeshRenderer>().enabled = false;
-
-        waterContainer.GetComponent<waterEffectController>().enabled = false;
-        waterContainer.GetComponent<BoxCollider>().enabled = false;
-    }
-
-}
