--- DH2650/Assets/Scripts/DisableWater.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d4c7a69a8b4f6f44a9bd082db6da40d8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
