--- DH2650/Assets/Scripts/EnableWater.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0c5860f1258a33447a0217aa1cf072de
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
