+++ DH2650/Assets/Scripts/Interactables/DisableBoxCollider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c2c7d4b4c4d394148b211489ac4d352d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
