--- DH2650/Assets/Scripts/ResetLightning.cs
@@ -1,13 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class ResetLightning : MonoBehaviour
-{
-    public GameObject dirSkyLight;
-
-    private void OnTriggerEnter(Collider other)
-    {
-        dirSkyLight.GetComponent<Light>().enabled = true;
-    }
-}
