--- DH2650/Assets/Scripts/ResetLightning.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2667086a886e41e4dad02e829f0213ba
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
