+++ DH2650/ProjectSettings/TagManager.asset
@@ -19,7 +19,7 @@ TagManager:
   - UI
   - Ground
   - Player
-  - 
+  - NonGrappable
   - 
   - 
   -
